Double quotes for styles

+++ .stylelintrc.js
@@ -2,7 +2,7 @@ module.exports = {
   'extends': 'stylelint-config-standard',
   'rules': {
     'no-empty-source': null,
-    'string-quotes': 'single',
+    'string-quotes': 'double',
     'at-rule-no-unknown': [
       true,
       {
